refactor: Split functions into different files
