+++ product-service/handlers/index.ts
@@ -1,0 +1,2 @@
+export { getProductsList } from './getProductsList';
+export { getProductById } from './getProductById';
